refactor: split .hpp implementations into .cpp, add clangd/compile_commands setup
Converts the codebase from a single-TU design (all code in headers, legacy
of an abandoned C++20 modules experiment) to a conventional multi-TU layout.
Each module now has a slim .hpp with declarations and a .cpp with definitions.

- 21 headers split into declaration+implementation pairs
- CMakeLists.txt: explicit CHRONOS_SOURCES list, CMAKE_EXPORT_COMPILE_COMMANDS,
  copy_compile_commands target copies compile_commands.json to repo root
- .clangd: points CompilationDatabase at repo root for zero-config editor setup
- Justfile: adds test, debug, clean targets
- README: adds Editor setup section
- ARCHITECTURE.md: updates layer diagram and removes single-TU design note
- .gitignore: adds compile_commands.json (machine-specific, generated on build)
