+++ .clangd
@@ -1,0 +1,4 @@
+CompileFlags:
+  CompilationDatabase: .
+Diagnostics:
+  UnusedIncludes: Strict
