+++ .gitignore
@@ -3,3 +3,4 @@ build_*/
 build-*/
 stopwatch-*.txt
 CMakeUserPresets.json
+compile_commands.json
