+++ src/painting.cpp
@@ -1,0 +1,268 @@
+#include "painting.hpp"
+#include <windows.h>
+#include <chrono>
+#include <filesystem>
+#include <format>
+#include <optional>
+#include <string>
+#include <vector>
+#include "assert.hpp"
+#include "actions.hpp"
+#include "encoding.hpp"
+#include "app.hpp"
+#include "formatting.hpp"
+#include "layout.hpp"
+#include "paint_ctx.hpp"
+#include "painting_analog.hpp"
+#include "painting_timer.hpp"
+#include "theme.hpp"
+
+// ─── Paint sub-functions ──────────────────────────────────────────────────────
+int paint_bar(HDC hdc, int cw, PaintCtx& ctx) {
+    auto& layout = ctx.layout;
+    RECT bar{0, 0, cw, layout.bar_h};
+    FillRect(hdc, &bar, ctx.res.brBar);
+
+    SelectObject(hdc, ctx.res.fontSm);
+    int by = (layout.bar_h - layout.btn_h) / 2;
+    int bx = (cw - (layout.w_pin + layout.w_clk + layout.w_sw + layout.w_tmr + layout.w_alm + layout.w_set + 5 * layout.bar_gap)) / 2;
+    btn(hdc, {bx, by, bx + layout.w_pin, by + layout.btn_h}, ctx.app.topmost, L"Pin", A_TOPMOST, ctx);
+    bx += layout.w_pin + layout.bar_gap;
+    btn(hdc, {bx, by, bx + layout.w_clk, by + layout.btn_h}, ctx.app.show_clk, L"Clock", A_SHOW_CLK, ctx);
+    bx += layout.w_clk + layout.bar_gap;
+    btn(hdc, {bx, by, bx + layout.w_sw, by + layout.btn_h}, ctx.app.show_sw, L"Stopwatch", A_SHOW_SW, ctx);
+    bx += layout.w_sw + layout.bar_gap;
+    btn(hdc, {bx, by, bx + layout.w_tmr, by + layout.btn_h}, ctx.app.show_tmr, L"Timers", A_SHOW_TMR, ctx);
+    bx += layout.w_tmr + layout.bar_gap;
+    btn(hdc, {bx, by, bx + layout.w_alm, by + layout.btn_h}, ctx.app.show_alarms, L"Alarms", A_SHOW_ALARMS, ctx);
+    bx += layout.w_alm + layout.bar_gap;
+    btn(hdc, {bx, by, bx + layout.w_set, by + layout.btn_h}, false, L"\u2699", A_SETTINGS, ctx);
+    return layout.bar_h;
+}
+
+int paint_clock(HDC hdc, int cw, int y, PaintCtx& ctx) {
+    auto& layout = ctx.layout;
+    divider(hdc, y, cw, ctx);
+    SYSTEMTIME st;
+    GetLocalTime(&st);
+
+    if (ctx.app.clock_view == ClockView::Analog) {
+        RECT area{0, y, cw, y + layout.analog_clk_h};
+        draw_analog_clock(hdc, area, ctx.app.analog_style, ctx.theme,
+                          layout.dpi, st.wHour, st.wMinute, st.wSecond);
+        ctx.btns.push_back({area, A_CLK_CYCLE});
+        return y + layout.analog_clk_h;
+    }
+
+    std::wstring buf;
+    switch (ctx.app.clock_view) {
+    case ClockView::H24_HMS:
+        buf = std::format(L"{:02}:{:02}:{:02}", st.wHour, st.wMinute, st.wSecond);
+        break;
+    case ClockView::H24_HM:
+        buf = std::format(L"{:02}:{:02}", st.wHour, st.wMinute);
+        break;
+    case ClockView::H12_HMS: {
+        int h = st.wHour % 12;
+        if (h == 0) h = 12;
+        buf = std::format(L"{}:{:02}:{:02} {}", h, st.wMinute, st.wSecond, st.wHour < 12 ? L"AM" : L"PM");
+        break;
+    }
+    case ClockView::H12_HM: {
+        int h = st.wHour % 12;
+        if (h == 0) h = 12;
+        buf = std::format(L"{}:{:02} {}", h, st.wMinute, st.wHour < 12 ? L"AM" : L"PM");
+        break;
+    }
+    default: CHRONOS_UNREACHABLE();
+    }
+    SelectObject(hdc, ctx.res.fontBig);
+    SetTextColor(hdc, ctx.theme.text);
+    RECT tr{0, y, cw, y + layout.clk_h};
+    DrawTextW(hdc, buf.c_str(), -1, &tr, DT_CENTER | DT_VCENTER | DT_SINGLELINE);
+    ctx.btns.push_back({tr, A_CLK_CYCLE});
+    return y + layout.clk_h;
+}
+
+int paint_stopwatch(HDC hdc, int cw, int y, PaintCtx& ctx, std::chrono::steady_clock::time_point now) {
+    auto& layout = ctx.layout;
+    auto& th = ctx.theme;
+    divider(hdc, y, cw, ctx);
+    auto elap = ctx.app.sw.elapsed(now);
+    std::wstring etime = format_stopwatch_display(elap);
+
+    SelectObject(hdc, ctx.res.fontBig);
+    SetTextColor(hdc, th.text);
+    RECT tr{0, y + layout.dpi_scale(4), cw, y + layout.dpi_scale(44)};
+    DrawTextW(hdc, etime.c_str(), -1, &tr, DT_CENTER | DT_VCENTER | DT_SINGLELINE);
+
+    SelectObject(hdc, ctx.res.fontSm);
+    bool running = ctx.app.sw.is_running();
+    int bw = layout.dpi_scale(76), gap = layout.dpi_scale(6), bh = layout.dpi_scale(28);
+    int bx0 = (cw - 3 * bw - 2 * gap) / 2;
+    int by0 = y + layout.dpi_scale(46);
+    btn(hdc, {bx0, by0, bx0 + bw, by0 + bh}, running, running ? L"Stop" : L"Start", A_SW_START, ctx);
+    btn(hdc, {bx0 + bw + gap, by0, bx0 + 2 * bw + gap, by0 + bh}, false, L"Lap", A_SW_LAP, ctx);
+    btn(hdc, {bx0 + 2 * (bw + gap), by0, bx0 + 3 * bw + 2 * gap, by0 + bh}, false, L"Reset", A_SW_RESET, ctx);
+
+    auto& laps = ctx.app.sw.laps();
+    auto info = laps.empty() ? std::wstring(L"\u2014")
+                             : std::format(L"Lap {}  \u2014  {}", laps.size(), format_stopwatch_short(laps.back()));
+    SetTextColor(hdc, th.dim);
+    RECT ir{0, by0 + bh + layout.dpi_scale(4), cw, by0 + bh + layout.dpi_scale(22)};
+    DrawTextW(hdc, info.c_str(), -1, &ir, DT_CENTER | DT_VCENTER | DT_SINGLELINE);
+
+    bool has_file = !ctx.app.sw_lap_file.empty();
+    int gbw = layout.dpi_scale(100), gbh = layout.dpi_scale(18);
+    auto lap_label = ctx.app.lap_write_failed ? L"Get Laps (!)" : L"Get Laps";
+    auto lap_col = ctx.app.lap_write_failed ? th.expire : has_file ? th.btn : th.dim;
+    int laps_top = y + layout.sw_h - gbh;
+    btn(hdc, {(cw - gbw) / 2, laps_top, (cw + gbw) / 2, y + layout.sw_h},
+        false, lap_label, has_file ? A_SW_GET : 0, ctx, lap_col);
+    return y + layout.sw_h;
+}
+
+int paint_alarms(HDC hdc, int cw, int y, PaintCtx& ctx) {
+    auto& layout = ctx.layout;
+    auto& th = ctx.theme;
+    divider(hdc, y, cw, ctx);
+
+    int pad = layout.dpi_scale(10);
+    int bw  = layout.dpi_scale(56);
+    int bh  = layout.dpi_scale(24);
+    int by0 = y + (layout.alarm_header_h - bh) / 2;
+
+    SelectObject(hdc, ctx.res.fontSm);
+    SetTextColor(hdc, th.text);
+    RECT title_r{pad, y, cw - bw - 2 * pad, y + layout.alarm_header_h};
+    DrawTextW(hdc, L"Alarms", -1, &title_r, DT_LEFT | DT_VCENTER | DT_SINGLELINE);
+
+    btn(hdc, {cw - bw - pad, by0, cw - pad, by0 + bh}, false, L"+ Add", A_ALARM_ADD, ctx);
+
+    int row_y = y + layout.alarm_header_h;
+    int del_w = layout.dpi_scale(36);
+    int tog_w = layout.dpi_scale(50);
+
+    HBRUSH stripe = CreateSolidBrush(th.bar);
+    for (int i = 0; i < (int)ctx.app.alarms.size(); ++i) {
+        const auto& a = ctx.app.alarms[i];
+
+        RECT row_bg{0, row_y, cw, row_y + layout.alarm_row_h};
+        if (i % 2 == 1)
+            FillRect(hdc, &row_bg, stripe);
+
+        // Build schedule string
+        std::wstring sched;
+        if (a.schedule == AlarmSchedule::Days) {
+            const wchar_t* day_names[] = {L"Mo", L"Tu", L"We", L"Th", L"Fr", L"Sa", L"Su"};
+            for (int d = 0; d < 7; ++d)
+                if (a.days_mask & (1 << d)) {
+                    if (!sched.empty()) sched += L' ';
+                    sched += day_names[d];
+                }
+            if (sched.empty()) sched = L"(no days)";
+        } else {
+            sched = std::format(L"{:04}-{:02}-{:02}", a.date_year, a.date_month, a.date_day);
+        }
+        std::wstring label;
+        {
+            std::wstring name_w = utf8_to_wide(a.name);
+            if (!name_w.empty()) label = name_w + L"  ";
+        }
+        label += std::format(L"{:02}:{:02}  {}", a.hour, a.minute, sched);
+
+        int text_x = pad;
+        int row_mid = row_y + layout.alarm_row_h / 2;
+        int text_h  = layout.dpi_scale(18);
+        int buttons_right = cw - pad;
+        int del_x  = buttons_right - del_w;
+        int tog_x  = del_x - layout.dpi_scale(6) - tog_w;
+
+        RECT lr{text_x, row_mid - text_h / 2, tog_x - layout.dpi_scale(6), row_mid + text_h / 2};
+        SetTextColor(hdc, a.enabled ? th.text : th.dim);
+        DrawTextW(hdc, label.c_str(), -1, &lr, DT_LEFT | DT_VCENTER | DT_SINGLELINE | DT_END_ELLIPSIS);
+
+        int btn_top = row_mid - bh / 2;
+        btn(hdc, {tog_x, btn_top, tog_x + tog_w, btn_top + bh},
+            a.enabled, a.enabled ? L"On" : L"Off", A_ALARM_TOGGLE + i, ctx);
+        btn(hdc, {del_x, btn_top, del_x + del_w, btn_top + bh},
+            false, L"Del", A_ALARM_DEL + i, ctx);
+
+        row_y += layout.alarm_row_h;
+    }
+    DeleteObject(stripe);
+
+    if (ctx.app.alarms.empty()) {
+        SetTextColor(hdc, th.dim);
+        RECT er{pad, row_y, cw - pad, row_y + layout.alarm_row_h};
+        DrawTextW(hdc, L"No alarms set", -1, &er, DT_CENTER | DT_VCENTER | DT_SINGLELINE);
+        row_y += layout.alarm_row_h;
+    }
+
+    return row_y;
+}
+
+void paint_help(HDC hdc, int cw, int y_bottom, PaintCtx& ctx) {
+    auto& layout = ctx.layout;
+    RECT cr{0, layout.bar_h, cw, y_bottom > layout.bar_h ? y_bottom : layout.bar_h + layout.dpi_scale(200)};
+    FillRect(hdc, &cr, ctx.res.brHelp);
+
+    SelectObject(hdc, ctx.res.fontSm);
+    SetTextColor(hdc, ctx.theme.text);
+
+    struct {
+        const wchar_t* key;
+        const wchar_t* desc;
+    } shortcuts[] = {
+        {L"Space", L"Start/Stop stopwatch or first timer"},
+        {L"L", L"Record lap"},
+        {L"E", L"Open exported laps file"},
+        {L"C", L"Copy laps to clipboard"},
+        {L"R", L"Reset stopwatch or first timer"},
+        {L"P", L"Toggle Pomodoro on first timer"},
+        {L"N", L"Skip to next Pomodoro phase"},
+        {L"T", L"Toggle always-on-top"},
+        {L"D", L"Cycle theme: Auto \u2192 Dark \u2192 Light"},
+        {L"1-9", L"Start/Stop timer 1-9"},
+        {L"Shift+1-9", L"Reset timer 1-9"},
+        {L"Shift+R", L"Reset all timers"},
+        {L"+ / =", L"Add a timer slot"},
+        {L"-", L"Remove last timer slot"},
+        {L"H / ?", L"Toggle this help"},
+        {L"Ctrl+Shift+Space", ctx.global_hotkey_ok ? L"Global start/stop (any app)"
+                                                    : L"Global start/stop (unavailable)"},
+        {L"", L""},
+        {L"Scroll", L"Adjust timer value (H/M/S)"},
+        {L"Click clock", L"Cycle clock format"},
+        {L"Dbl-click", L"Edit timer label"},
+        {L"Right-click", L"Timer presets / Pomodoro (untouched)"},
+    };
+
+    int line_h = layout.dpi_scale(20);
+    int pad = layout.dpi_scale(12);
+    int sy = layout.bar_h + pad;
+    for (auto& [key, desc] : shortcuts) {
+        auto line = std::format(L"  {}  \u2014  {}", key, desc);
+        RECT lr{pad, sy, cw - pad, sy + line_h};
+        DrawTextW(hdc, line.c_str(), -1, &lr, DT_LEFT | DT_VCENTER | DT_SINGLELINE);
+        sy += line_h;
+    }
+}
+
+// ─── Paint dispatcher ─────────────────────────────────────────────────────────
+void paint_all(HDC hdc, int cw, int ch, PaintCtx& ctx) {
+    ctx.btns.clear();
+    SetBkMode(hdc, TRANSPARENT);
+
+    RECT all{0, 0, cw, ch};
+    FillRect(hdc, &all, ctx.res.brBg);
+
+    ctx.now = std::chrono::steady_clock::now();
+    auto now = ctx.now;
+    int y = paint_bar(hdc, cw, ctx);
+    if (ctx.app.show_clk) y = paint_clock(hdc, cw, y, ctx);
+    if (ctx.app.show_sw) y = paint_stopwatch(hdc, cw, y, ctx, now);
+    if (ctx.app.show_tmr) y = paint_timers(hdc, cw, y, ctx, now);
+    if (ctx.app.show_alarms) y = paint_alarms(hdc, cw, y, ctx);
+    if (ctx.app.show_help) paint_help(hdc, cw, y, ctx);
+}
